fixed clock.c

+++ unixsyscall/clock/clock/src/clock/clock.c
@@ -29,7 +29,7 @@ int main(void){
   end = clock(); /* clockで終了時刻(プロセッサ時間単位)を取得し, endに格納. */
 
   /* 経過時間の計算 */
-  elapsed = (double)(end - start) / CLOCKS_PER_SEC; /* endからstartを引いてCLOCKS_PER_SECで割ると秒単位の経過時間が算出できるので, それをelapsedに格納. */
+  elapsed = (double)(end - start) / (double)CLOCKS_PER_SEC; /* endからstartを引いてCLOCKS_PER_SECで割ると秒単位の経過時間が算出できるので, それをelapsedに格納. */
 
   /* 経過時間の出力 */
   printf("elapsed = %lf\n", elapsed); /* printfで経過時間elapsedを出力. */
